Plot_Star_Col-Col.py:
updated to include the full sample for the background
colour-colour plot

discovered that SG_Flag is missing ~10% of the objects
- something to improve for DR5
I don't currently understand why this is, but I've removed
the SGFlag cut from the study now and just use
the PSF res cats to define the sample

included a 3 sigma criteria to include the noise on J-K
in our fraction of contaminants.  The J-K error around
g-i~0.5 can be as high as ~1 magnitude

Finally I removed some of the plot features which were
essential for diagnostics, but made for a hard-to-read
final plot for publication.   Simplified final plot but
kept lines in code for uncommenting in future
studies as these extensions are really key for the
super users!